requirements.txt
streamlit
torch
numpy
pandas
tqdm
fair-esm
